Refactor MCP JVM DTO parser
